Minor fixes for EthernetFrame

+++ protsim/src/main/java/fi/utu/protproc/group3/protocols/EthernetFrame.java
@@ -5,7 +5,7 @@ package fi.utu.protproc.group3.protocols;
  * Represents an 802.3 ethernet frame.
  */
 public interface EthernetFrame {
-    static EthernetFrame create(byte[] destination, byte[] source, short type, byte[] payload) {
+    static EthernetFrame create(byte[] destination, byte[] source, int type, byte[] payload) {
         throw new UnsupportedOperationException();
     }
 
@@ -13,12 +13,12 @@ public interface EthernetFrame {
         throw new UnsupportedOperationException();
     }
 
-    short TYPE_IPV4 = (short) 0x8000;
-    short TYPE_IPV6 = (short) 0x86dd;
+    int TYPE_IPV4 = 0x8000;
+    int TYPE_IPV6 = 0x86dd;
 
     byte[] getDestination();
     byte[] getSource();
-    short getType();
+    int getType();
     byte[] getPayload();
 
     /**
